test_remove_redundant_entries_oeq_class
