+++ datafusion/physical-expr/src/equivalence/mod.rs
@@ -375,244 +375,6 @@ mod tests {
     }
 
     #[test]
-    fn test_remove_redundant_entries_oeq_class() -> Result<()> {
-        let schema = create_test_schema()?;
-        let col_a = &col("a", &schema)?;
-        let col_b = &col("b", &schema)?;
-        let col_c = &col("c", &schema)?;
-        let col_d = &col("d", &schema)?;
-        let col_e = &col("e", &schema)?;
-
-        let option_asc = SortOptions {
-            descending: false,
-            nulls_first: false,
-        };
-        let option_desc = SortOptions {
-            descending: true,
-            nulls_first: true,
-        };
-
-        // First entry in the tuple is the given orderings for the table
-        // Second entry is the simplest version of the given orderings that is functionally equivalent.
-        let test_cases = vec![
-            // ------- TEST CASE 1 ---------
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [a ASC, b ASC]
-                    vec![(col_a, option_asc), (col_b, option_asc)],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [a ASC, b ASC]
-                    vec![(col_a, option_asc), (col_b, option_asc)],
-                ],
-            ),
-            // ------- TEST CASE 2 ---------
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [a ASC, b ASC]
-                    vec![(col_a, option_asc), (col_b, option_asc)],
-                    // [a ASC, b ASC, c ASC]
-                    vec![
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [a ASC, b ASC, c ASC]
-                    vec![
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                ],
-            ),
-            // ------- TEST CASE 3 ---------
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [a ASC, b DESC]
-                    vec![(col_a, option_asc), (col_b, option_desc)],
-                    // [a ASC]
-                    vec![(col_a, option_asc)],
-                    // [a ASC, c ASC]
-                    vec![(col_a, option_asc), (col_c, option_asc)],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [a ASC, b DESC]
-                    vec![(col_a, option_asc), (col_b, option_desc)],
-                    // [a ASC, c ASC]
-                    vec![(col_a, option_asc), (col_c, option_asc)],
-                ],
-            ),
-            // ------- TEST CASE 4 ---------
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [a ASC, b ASC]
-                    vec![(col_a, option_asc), (col_b, option_asc)],
-                    // [a ASC, b ASC, c ASC]
-                    vec![
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                    // [a ASC]
-                    vec![(col_a, option_asc)],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [a ASC, b ASC, c ASC]
-                    vec![
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                ],
-            ),
-            // ------- TEST CASE 5 ---------
-            // Empty ordering
-            (
-                vec![vec![]],
-                // No ordering in the state (empty ordering is ignored).
-                vec![],
-            ),
-            // ------- TEST CASE 6 ---------
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [a ASC, b ASC]
-                    vec![(col_a, option_asc), (col_b, option_asc)],
-                    // [b ASC]
-                    vec![(col_b, option_asc)],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [a ASC]
-                    vec![(col_a, option_asc)],
-                    // [b ASC]
-                    vec![(col_b, option_asc)],
-                ],
-            ),
-            // ------- TEST CASE 7 ---------
-            // b, a
-            // c, a
-            // d, b, c
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [b ASC, a ASC]
-                    vec![(col_b, option_asc), (col_a, option_asc)],
-                    // [c ASC, a ASC]
-                    vec![(col_c, option_asc), (col_a, option_asc)],
-                    // [d ASC, b ASC, c ASC]
-                    vec![
-                        (col_d, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [b ASC, a ASC]
-                    vec![(col_b, option_asc), (col_a, option_asc)],
-                    // [c ASC, a ASC]
-                    vec![(col_c, option_asc), (col_a, option_asc)],
-                    // [d ASC]
-                    vec![(col_d, option_asc)],
-                ],
-            ),
-            // ------- TEST CASE 8 ---------
-            // b, e
-            // c, a
-            // d, b, e, c, a
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [b ASC, e ASC]
-                    vec![(col_b, option_asc), (col_e, option_asc)],
-                    // [c ASC, a ASC]
-                    vec![(col_c, option_asc), (col_a, option_asc)],
-                    // [d ASC, b ASC, e ASC, c ASC, a ASC]
-                    vec![
-                        (col_d, option_asc),
-                        (col_b, option_asc),
-                        (col_e, option_asc),
-                        (col_c, option_asc),
-                        (col_a, option_asc),
-                    ],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [b ASC, e ASC]
-                    vec![(col_b, option_asc), (col_e, option_asc)],
-                    // [c ASC, a ASC]
-                    vec![(col_c, option_asc), (col_a, option_asc)],
-                    // [d ASC]
-                    vec![(col_d, option_asc)],
-                ],
-            ),
-            // ------- TEST CASE 9 ---------
-            // b
-            // a, b, c
-            // d, a, b
-            (
-                // ORDERINGS GIVEN
-                vec![
-                    // [b ASC]
-                    vec![(col_b, option_asc)],
-                    // [a ASC, b ASC, c ASC]
-                    vec![
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                    // [d ASC, a ASC, b ASC]
-                    vec![
-                        (col_d, option_asc),
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                    ],
-                ],
-                // EXPECTED orderings that is succinct.
-                vec![
-                    // [b ASC]
-                    vec![(col_b, option_asc)],
-                    // [a ASC, b ASC, c ASC]
-                    vec![
-                        (col_a, option_asc),
-                        (col_b, option_asc),
-                        (col_c, option_asc),
-                    ],
-                    // [d ASC]
-                    vec![(col_d, option_asc)],
-                ],
-            ),
-        ];
-        for (orderings, expected) in test_cases {
-            let orderings = convert_to_orderings(&orderings);
-            let expected = convert_to_orderings(&expected);
-            let actual = OrderingEquivalenceClass::new(orderings.clone());
-            let actual = actual.orderings;
-            let err_msg = format!(
-                "orderings: {:?}, expected: {:?}, actual :{:?}",
-                orderings, expected, actual
-            );
-            assert_eq!(actual.len(), expected.len(), "{}", err_msg);
-            for elem in actual {
-                assert!(expected.contains(&elem), "{}", err_msg);
-            }
-        }
-
-        Ok(())
-    }
-
-    #[test]
     fn test_get_updated_right_ordering_equivalence_properties() -> Result<()> {
         let join_type = JoinType::Inner;
         // Join right child schema
